Submitted files from a user on discord
This is a open source flight simulator for windows. It will run on Windows XP and up. It can use D3D9, Open GL1, or Open GL2. It isn't graphically intensive so a Pentium 4 machine with it's iGPU should be able to run it. Just have Windows XP, 512mb of RAM, 35MB of disc space, and a CPU faster than 950MHz